Revert "start_time issue fixed"
This reverts commit 0eb10c08a8c0a016b26cd1ab29af527be330d206.

+++ class/omlBasePackage/OMLBase.java
@@ -256,7 +256,7 @@ public class OMLBase {
 	private synchronized void create_head(){
 		// Take current time
 		head_time = System.currentTimeMillis();
-		int lcl_head_time = (int)(head_time/1000L);
+		float lcl_head_time = head_time/1000L;
 		String time = String.valueOf(lcl_head_time);
 		 
 		header.append("protocol: 1\n");
